backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
paramiko==3.4.0
influxdb-client==1.35.0
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.0
imbalanced-learn==0.11.0
